Lesson 3 snippets update
