+++ tutorials/nservicebus-step-by-step/3-multiple-endpoints/Snippets/Core_9/Lesson3.cs
@@ -1,0 +1,106 @@
+﻿using Microsoft.Extensions.Hosting;
+using NServiceBus;
+using System;
+using System.Threading.Tasks;
+
+namespace Core_9;
+
+class Snippets
+{
+    [System.Diagnostics.CodeAnalysis.SuppressMessage("NServiceBus.Code", "NSB0002:Forward the 'CancellationToken' property of the context parameter to methods", Justification = "Parameters only both included for ease of documentation")]
+    public async Task SendTypes(IMessageSession messageSession, IMessageHandlerContext context)
+    {
+        var command = new object();
+
+        #region SendLocal
+        // From endpoint startup code
+        await messageSession.SendLocal(command);
+
+        // From a message handler
+        await context.SendLocal(command);
+        #endregion
+
+        #region Send
+        // From endpoint startup code
+        await messageSession.Send(command);
+
+        // From a message handler
+        await context.Send(command);
+        #endregion
+
+        #region SendDestination
+        // Not recommended, most of the time
+        await messageSession.Send("Destination", command);
+
+        // On the IMessageHandlerContext too, but still not recommended
+        await context.Send("Destination", command);
+
+        #endregion
+
+    }
+
+    void ShowRouting(EndpointConfiguration endpointConfiguration)
+    {
+        #region RoutingSettings
+        var transport = endpointConfiguration.UseTransport<LearningTransport>();
+        // Returns a RoutingSettings<LearningTransport>
+        var routing = transport.Routing();
+        #endregion
+
+        #region RouteToEndpoint
+        // Specify the routing for a specific type
+        routing.RouteToEndpoint(typeof(DoSomething), "SomeEndpoint");
+
+        // Specify the routing for all messages in an assembly
+        routing.RouteToEndpoint(typeof(DoSomething).Assembly, "SomeEndpoint");
+
+        // Specify the routing for all messages in a given assembly and namespace
+        routing.RouteToEndpoint(typeof(DoSomething).Assembly, "Specific.Namespace", "SomeEndpoint");
+
+        #endregion
+    }
+
+    class DoSomething { }
+    class PlaceOrder { }
+
+    void ExerciseConfig()
+    {
+        #region EndpointDifferences
+        Console.Title = "Sales";
+
+        var endpointConfiguration = new EndpointConfiguration("Sales");
+        #endregion
+
+        var transport = endpointConfiguration.UseTransport<LearningTransport>();
+
+        #region AddingRouting
+        var routing = transport.Routing();
+        routing.RouteToEndpoint(typeof(PlaceOrder), "Sales");
+        #endregion
+
+    }
+
+    #region SalesConsoleApp
+
+    class Program
+    {
+        static async Task Main(string[] args)
+        {
+            Console.Title = "Sales";
+
+            var builder = Host.CreateApplicationBuilder(args);
+
+            var endpointConfiguration = new EndpointConfiguration("Sales");
+
+            endpointConfiguration.UseSerialization<SystemJsonSerializer>();
+
+            endpointConfiguration.UseTransport<LearningTransport>();
+
+            builder.UseNServiceBus(endpointConfiguration);
+
+            await builder.Build().RunAsync();
+        }
+    }
+
+    #endregion
+}
